+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # cit-context ChangeLog
 
+### Fixed
+- Add `appContextMap` and `documentLoader` to named rollup exports.
+
 ## 1.0.0 - 2020-09-09
 
 Initial version.